common: optimize paint data size.
packing the data fields with the appropriate size.

96 -> 24 (-9 bytes per one paint)

+++ src/lib/tvgFill.h
@@ -55,7 +55,7 @@ struct Fill::Impl
     uint32_t cnt = 0;
     FillSpread spread;
     DuplicateMethod<Fill>* dup = nullptr;
-    uint32_t id;
+    uint8_t id;
 
     ~Impl()
     {
